Move build logic to lib/build
